move withTag to own map to fix gen components.ts
