+++ src/layout/FileUpload/Summary/AttachmentWithTagSummaryComponent.test.tsx
@@ -7,7 +7,7 @@ import { AttachmentSummaryComponent } from 'src/layout/FileUpload/Summary/Attach
 import { renderWithProviders } from 'src/testUtils';
 import { useResolvedNode } from 'src/utils/layout/ExprContext';
 import type { ExprUnresolved } from 'src/features/expressions/types';
-import type { ILayoutCompFileUploadWithTag } from 'src/layout/FileUpload/FileUploadWithTag/types';
+import type { ILayoutCompFileUploadWithTag } from 'src/layout/FileUploadWithTag/types';
 import type { RootState } from 'src/redux/store';
 import type { LayoutNodeFromType } from 'src/utils/layout/hierarchy.types';
